configk: write output in config file format
Add new command-line option '-C' to write output in a config file
format. User can redirect such output to a file and use it.

Add support to supply value while enabling an option with '-e'.

Add support to read 'SRCARCH' environment variable when it's not
given with '-a' command-line option.