README: update documentation with new options
Add a 'Quick Start' section at the top, it'll help user to try
configk commands and explore its functions.

List -C --config option in the README document. It writes the
output in a config file format. User can redirect it to a file
and use it.

Minor updates to validate_option function to handle commented OR
"is not set" options from a give configuration file. Need to break
the validate_option function into two, one to set option values
and another to validate values.

Skip toggle operation for non-tristate type options.